Make module not sks-specific
Turns out we'll need to use the module as a generic one after all, so it won't have the auto-injected locals and variables. Now it's a completely generic and re-usable module outside of our sks-specific setup.

Which also means that we can go back to an open source license and make the repo public